Break server security test files into smaller units
Change-type: patch
