+++ apps/server/card-loader.js
@@ -36,7 +36,7 @@ module.exports = async (context, jellyfish, worker, session) => {
 	await Bluebird.each([
 		// Roles
 		await loadCard('contrib/role-user-community.json'),
-        await loadCard('contrib/role-user-operator.json'),
+		await loadCard('contrib/role-user-operator.json'),
 		await loadCard('contrib/role-user-guest.json'),
 		!environment.isProduction() && await loadCard('contrib/role-user-test.json'),
 		await loadCard('contrib/role-user-external-support.json'),
