+++ test/e2e/server/security/helpers.js
@@ -1,0 +1,19 @@
+/*
+ * Copyright (C) Balena.io - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited.
+ * Proprietary and confidential.
+ */
+
+const uuid = require('uuid/v4')
+
+const createUserDetails = () => {
+	return {
+		username: uuid(),
+		email: `${uuid()}@example.com`,
+		password: 'foobarbaz'
+	}
+}
+
+module.exports = {
+	createUserDetails
+}
